Temporary workaround for context sensitive dictation in Citrix Viewer in Talon beta.
Doesn't quite work perfectly (hits 150 ms timeout if there is nothing to copy).

Basic issue is that Citrix Viewer gets confused when clip.clear() removes all pasteboard types, so we set a different clipboard placeholder.

On my setup, time to get the clipboard is in the range of 50-100 ms; may need to adjust the range if your setup is slower.