Add spoken commands for bitwise and/or assignment (#958)
As far as I can tell, these actions were never given spoken forms. I am
not sure if there is a reason for this, so I wanted to add some in case
it was an error.

I made "wise" optional in the command so it does not get too long, but
the square brackets can be removed if that does not seem necessary. If
this is useful in reducing command length, maybe similar commands could
be formatted to make the "wise" in "bitwise" optional.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>