Add some more commands for handling invitations and notifications.
Also press the down arrow after opening the notifications popover so a notification is selected.

May have to do this with accessibility as it does not seem entirely reliable to do so via the keyboard.

Note: dismiss/dismiss all keyboard equivalents currently require a beta of Fantastical 3.7.6; it should be in the final release for everyone.