Fixes for OneNote "collapse this".
Collapse/restore the ribbon in OneNote immediately if we only expanded it to check the zoom level.  Not only does this prevent the View ribbon tab from remaining visible, it helps with a race in which the ribbon animates after we've matched a selection location, since the original location will be valid.

Move the mouse out of the way first, so that we do not match an outline handle from the original mouse location.

Shorten the arbitrary interval we wait between moving the mouse and double clicking as it does not seem to be necessary to make it that long. (Suspect this will be highly dependent on hardware and page complexity...)

If you are doing a lot of "collapsing this", it is most reliable/fastest to start with the ribbon visible and "zoom reset" (100%).