Hack to improve reliability of OneNote "collapse this".
Try to move the mouse pointer out of the way as early as practical.

Still need an extra delay since otherwise we sometimes match against a stale outline handle and collapse the wrong thing.

A possibility might be to wait until there is no matching outline handle before we even try to move the mouse. This might help with other race conditions as well.