Migrate php operators (#1811)
Migrate php operators. I believe this is the final operators migration.
I added some missing operators. There are some further operators that
might be worth adding, but that might be best left to pull requests from
people who understand the language.

I wonder if adding "is loosely equal" and "is loosely not equal" as
aliases might be better than deprecation.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>